Rename src package to stravavis
