+++ setup.cfg
@@ -48,4 +48,4 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    stravavis = strava_py.cli:main
+    stravavis = stravavis.cli:main
