+++ src/stravavis/__main__.py
@@ -1,4 +1,4 @@
-from strava_py import cli
+from stravavis import cli
 
 if __name__ == "__main__":
     cli.main()
