+++ src/stravavis/plot_facets.py
