+++ src/stravavis/plot_map.py
